Cambios en el script de la clase 1
